+++ examples/cursor.rs
@@ -2,6 +2,7 @@
 
 use simple_logger::SimpleLogger;
 use winit::{
+    dpi::LogicalSize,
     event::{ElementState, Event, KeyboardInput, WindowEvent},
     event_loop::EventLoop,
     window::{CursorIcon, WindowBuilder},
@@ -10,10 +11,12 @@ use winit::{
 fn main() {
     SimpleLogger::new().init().unwrap();
     let event_loop = EventLoop::new();
-
-    let window = WindowBuilder::new().build(&event_loop).unwrap();
+    let window = WindowBuilder::new()
+        .with_inner_size(LogicalSize::new(800.0, 600.0))
+        .with_centered_position()
+        .build(&event_loop)
+        .unwrap();
     window.set_title("A fantastic window!");
-
     let mut cursor_idx = 0;
 
     event_loop.run(move |event, _, control_flow| {